feat: Add Hudi Spark master and worker services to docker-compose.yml
The code changes add the Hudi Spark master and worker services to the docker-compose.yml file. The master service uses the latest bitnami/spark image and is configured with volumes and ports. The worker service is also based on the latest bitnami/spark image and is configured with environment variables and replicas.